+++ src/components/Toaster/hooks/useInToasterContext.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ToasterContext} from '../ToasterContext';
+import {ToasterContext} from '../Provider/ToasterContext';
 
 export function useInToasterContext() {
     return React.useContext(ToasterContext).isInitialized;
