+++ src/components/Toaster/hooks/useToaster.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ToasterContext} from '../ToasterContext';
+import {ToasterContext} from '../Provider/ToasterContext';
 import {ToasterPublicMethods} from '../types';
 
 export function useToaster(): ToasterPublicMethods {
